requirements.txt
# Core ML and Data Processing
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.74
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Federated Learning Infrastructure
kafka-python==2.0.2

# Evaluation and Metrics
pandas==2.0.3

# Logging and Monitoring
tqdm==4.66.1

# Utilities
python-dotenv==1.0.0